Backend/requirements.txt
Django==5.2.5
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0

# Optional (development)
pytest>=7.0.0
